M185, max_soft_inactive_time: added a command for setting an idle time after which fans, PSU and motors are turned off
M185 is only activated temporarily. Entering any command after issuing M185
(except M105, which Printrun sends automatically periodically if the watch
window is open) will disable this inactivation counter.

It is intended to be put as the last command after a print, when the fans
are probably blowing at full power and there might be nobody to attend
the machine. Then if you have M185 S900 as the last command, the printer
will turn off the fan and power down then PSU (if available) after 15
minutes. Should you return to the machine and issue commands manually,
the inactivation is disabled.